@@ -1,4 +1,5 @@
 import React from 'react';
+import ReactDOM from 'react-dom';
 import PropTypes from 'prop-types';
 import axios from 'axios';
 import CommonMultistepComponent from '../common/CommonMultistepComponent.jsx';
@@ -6,6 +7,7 @@ import ErrorViewComponent from '../common/ErrorViewComponent.jsx';
 import UploadingThrobber from '../common/UploadingThrobber.jsx';
 import FileEntryView from '../../EntryViews/FileEntryView.jsx';
 import StorageSelector from "../../Selectors/StorageSelector.jsx";
+import UploadNewVersionComponent from "./UploadNewVersionComponent.jsx";
 
 class TemplateUploadComponent extends CommonMultistepComponent {
     static propTypes = {
@@ -26,25 +28,31 @@ class TemplateUploadComponent extends CommonMultistepComponent {
             selectedStorage: firstStorage,
             fileId: null,
             uploading: false,
-            useExisting: true
+            useExisting: true,
+            uploadFileVersion: 1,
+            uploadFileName: null,
+            fileInputKey:0  //we use this to clear the file input component by forcing it to re-render when the value is changed
         };
 
         this.fileChange = this.fileChange.bind(this);
+        this.fileInput = React.createRef();
     }
 
     componentDidUpdate() {
         //override the super-class implementation as we don't want it here
     }
 
-    componentWillMount(){
+    UNSAFE_componentWillMount(){
         if(this.props.existingFileId===null) this.setState({useExisting: false});
     }
 
-    doUpload(file){
+    doUpload(){
+        const data = this.fileInput.current.files[0];
+        console.log("doUpload: file name is ", data.name);
         axios({
             method: "PUT",
             url: "/api/file/" + this.state.fileId + "/content",
-            data: file,
+            data: data,
             headers: {'Content-Type': 'application/octet-stream'}
         }).then(response=>{
             this.setState({uploading:false, hasFiles: true}, ()=>{
@@ -52,34 +60,50 @@ class TemplateUploadComponent extends CommonMultistepComponent {
             });
         }).catch(error=>{
             this.setState({uploading:false, error:error});
-        })
+        });
     }
 
-    createFile(filename){
+    createFile(){
         return new Promise((resolve,reject)=>{
-            this.setState({uploading: true, error: null},()=>{
-                const nowtime = new Date().toISOString();
+            if(this.state.uploadFileName===null){
+                reject("uploadFileName not set")
+            } else {
+                this.setState({uploading: true, error: null}, () => {
+                    const nowtime = new Date().toISOString();
 
-                axios({
-                    method: "PUT",
-                    url: "/api/file",
-                    data: {
-                        storage: parseInt(this.state.selectedStorage),
-                        filepath: filename,
-                        version: 1,
-                        user: "",
-                        ctime: nowtime,
-                        mtime: nowtime,
-                        atime: nowtime,
-                        hasContent: false,
-                        hasLink: false
-                    }
-                }).then(response=>{
-                    this.setState({fileId: response.data.id}, ()=>resolve(response.data));
-                }).catch(error=>{
-                    this.setState({error: error, uploading: false}, ()=>reject(error));
-                })
-            });
+                    axios({
+                        method: "PUT",
+                        url: "/api/file",
+                        data: {
+                            storage: parseInt(this.state.selectedStorage),
+                            filepath: this.state.uploadFileName,
+                            version: this.state.uploadFileVersion,
+                            user: "",
+                            ctime: nowtime,
+                            mtime: nowtime,
+                            atime: nowtime,
+                            hasContent: false,
+                            hasLink: false
+                        }
+                    }).then(response => {
+                        this.setState({fileId: response.data.id}, () => resolve(response.data));
+                    }).catch(error => {
+                        console.log(error);
+                        console.log(error.response.data);
+                        console.log(error.response.data.hasOwnProperty("nextAvailableVersion"));
+                        if (error.response.data.hasOwnProperty("nextAvailableVersion")) {
+                            console.log("can create object at version ", error.response.data.nextAvailableVersion);
+                            this.setState({
+                                error: error,
+                                uploading: false,
+                                uploadFileVersion: error.response.data.nextAvailableVersion
+                            }, () => reject(error));
+                        } else {
+                            this.setState({error: error, uploading: false}, () => reject(error));
+                        }
+                    })
+                });
+            }
         });
     }
 
@@ -90,8 +114,10 @@ class TemplateUploadComponent extends CommonMultistepComponent {
             console.log("No files present");
             return;
         }
-        this.createFile(files[0].name).then(()=>
-            this.doUpload(files[0])
+        this.setState({uploadFileName: files[0].name}, ()=>
+            this.createFile().then(()=>
+                this.doUpload()
+            )
         );
     }
 
@@ -106,6 +132,11 @@ class TemplateUploadComponent extends CommonMultistepComponent {
         }
     }
 
+    storageSupportsVersions(){
+        const matchingStorageList = this.props.storages.filter((value,idx,arry)=>value===this.state.selectedStorage);
+        if(matchingStorageList.length===0) return false;
+        console.log("storageSupportsVersions: ", matchingStorageList[0].supportsVersions);
+    }
     render(){
         return <div>
             <h3>Template upload</h3>
@@ -123,12 +154,6 @@ class TemplateUploadComponent extends CommonMultistepComponent {
 
             <table style={{display: this.state.useExisting ? "none" : "inherit"}}>
                 <tbody>
-                {/*<tr>*/}
-                    {/*<td>Storage</td><td><select onChange={(event)=>this.setState({selectedStorage: event.target.value})}>*/}
-                    {/*{*/}
-                        {/*this.props.storages.map((strg,index)=><option key={index} value={strg.id}>{this.labelForStorage(strg)}</option>)*/}
-                    {/*}</select></td>*/}
-                {/*</tr>*/}
                 <tr>
                     <td>Storage</td>
                     <td><StorageSelector selectedStorage={this.state.selectedStorage}
@@ -140,12 +165,17 @@ class TemplateUploadComponent extends CommonMultistepComponent {
                 </tr>
                 <tr>
                     <td>File to upload</td>
-                    <td><input onChange={this.fileChange} type="file" id="template-upload-fileselector"/></td>
-                    <td><UploadingThrobber loading={this.state.uploading}/></td>
+                    <td><input onChange={this.fileChange} type="file" id="template-upload-fileselector" key={this.state.fileInputKey} ref={this.fileInput}/></td>
+                    <td><UploadingThrobber loading={this.state.uploading}/>{this.state.hasFiles? <p style={{color:"green"}}>File uploaded to version {this.state.uploadFileVersion}, please click "Next"</p> : <p/>}</td>
                 </tr>
                 </tbody>
             </table>
             <ErrorViewComponent error={this.state.error}/>
+            { (this.state.uploadFileVersion===1 || this.state.hasFiles) ? <p/> : <UploadNewVersionComponent
+                newVersionCb={()=>this.createFile().then(()=>this.doUpload())}
+                cancelCb={()=>this.setState({fileInputKey: this.state.fileInputKey+1, uploadFileVersion: 1, error:null})}/>
+            }
+
         </div>
     }
 }
