@@ -18,7 +18,7 @@ trait StorageDriver {
     * @param path [[String]] absolute path to write
     * @param dataStream [[java.io.FileInputStream]] to write from
     */
-  def writeDataToPath(path:String, dataStream:FileInputStream):Try[Unit]
+  def writeDataToPath(path:String, versionNumber: Int, dataStream:InputStream):Try[Unit]
 
   /**
     * Directly write a byte array to the given path (blocking)
@@ -26,28 +26,28 @@ trait StorageDriver {
     * @param data [[Array]] (of bytes) -  byte array to output
     * @return a Try indicating success or failure. If successful the Try has a unit value.
     */
-  def writeDataToPath(path:String, data:Array[Byte]):Try[Unit]
+  def writeDataToPath(path:String, versionNumber: Int, data:Array[Byte]):Try[Unit]
 
   /**
     * Delete the file at the given path (blocking)
     * @param path [[String]] absolute path to delete
     * @return [[Boolean]] indicating whether the file was deleted or not.
     */
-  def deleteFileAtPath(path:String):Boolean
+  def deleteFileAtPath(path:String, versionNumber: Int):Boolean
 
   /**
     * Get a relevant type of InputStream to read a file's data
     * @param path [[String]] Absolute path to open
     * @return [[java.io.InputStream]] subclass wrapped in a [[Try]]
     */
-  def getReadStream(path:String):Try[InputStream]
+  def getReadStream(path:String, versionNumber: Int):Try[InputStream]
 
   /**
     * Get a relevant type of OutputStream to write a file's data.  this may truncate the file.
     * @param path [[String]] Absolute path to open
     * @return [[java.io.OutputStream]] subclass wrapped in a [[Try]]
     */
-  def getWriteStream(path:String):Try[OutputStream]
+  def getWriteStream(path:String, versionNumber: Int):Try[OutputStream]
 
   /**
     * Get a Map of metadata relevant to the specified file.  The contents can vary between implementations, but should always
@@ -55,12 +55,18 @@ trait StorageDriver {
     * @param path [[String]] Absolute path to open
     * @return [[Map]] of [[Symbol]] -> [[String]] containing metadata about the given file.
     */
-  def getMetadata(path:String):Map[Symbol,String]
+  def getMetadata(path:String, versionNumber: Int):Map[Symbol,String]
 
   /**
     * Does the given path exist on this storage?
     * @param path
     * @return
     */
-  def pathExists(path:String):Boolean
+  def pathExists(path:String, versionNumber: Int):Boolean
+
+  /**
+    * Does this driver support or ignore version numbers?
+    * @return
+    */
+  def supportsVersions:Boolean
 }
