@@ -4,6 +4,7 @@ import org.joda.time.DateTime
 import slick.jdbc.PostgresProfile.api._
 import java.sql.Timestamp
 
+import akka.stream.Materializer
 import play.api.libs.functional.syntax._
 import play.api.libs.json.{JsPath, Reads, Writes}
 
@@ -27,6 +28,7 @@ trait StorageSerializer {
   /*https://www.playframework.com/documentation/2.5.x/ScalaJson*/
   implicit val storageWrites:Writes[StorageEntry] = (
     (JsPath \ "id").writeNullable[Int] and
+      (JsPath \ "nickname").writeNullable[String] and
       (JsPath \ "rootpath").writeNullable[String] and
       (JsPath \ "clientpath").writeNullable[String] and
       (JsPath \ "storageType").write[String] and
@@ -34,11 +36,14 @@ trait StorageSerializer {
       (JsPath \ "password").writeNullable[String] and
       (JsPath \ "host").writeNullable[String] and
       (JsPath \ "port").writeNullable[Int] and
+      (JsPath \ "device").writeNullable[String] and
+      (JsPath \ "supportsVersions").write[Boolean] and
       (JsPath \ "status").writeNullable[StorageStatus.Value]
     )(unlift(StorageEntry.unapply))
 
   implicit val storageReads:Reads[StorageEntry] = (
     (JsPath \ "id").readNullable[Int] and
+      (JsPath \ "nickname").readNullable[String] and
       (JsPath \ "rootpath").readNullable[String] and
       (JsPath \ "clientpath").readNullable[String] and
       (JsPath \ "storageType").read[String] and
@@ -46,18 +51,22 @@ trait StorageSerializer {
       (JsPath \ "password").readNullable[String] and
       (JsPath \ "host").readNullable[String] and
       (JsPath \ "port").readNullable[Int] and
+      (JsPath \ "device").readNullable[String] and
+      (JsPath \ "supportsVersions").read[Boolean] and
       (JsPath \ "status").readNullable[StorageStatus.Value]
     )(StorageEntry.apply _)
 }
 
 
-case class StorageEntry(id: Option[Int], rootpath: Option[String], clientpath: Option[String], storageType: String,
-                        user:Option[String], password:Option[String], host:Option[String], port:Option[Int], status:Option[StorageStatus.Value]) {
+case class StorageEntry(id: Option[Int], nickname:Option[String], rootpath: Option[String], clientpath: Option[String], storageType: String,
+                        user:Option[String], password:Option[String], host:Option[String], port:Option[Int], device:Option[String], supportsVersions: Boolean, status:Option[StorageStatus.Value]) {
 
-  def getStorageDriver:Option[StorageDriver] = {
+  def getStorageDriver(implicit mat:Materializer):Option[StorageDriver] = {
     val logger: Logger = Logger(this.getClass)
-    if(storageType=="Local"){
+    if(storageType=="Local") {
       Some(new PathStorage(this))
+    } else if(storageType=="ObjectMatrix") {
+      Some(new MatrixStoreDriver(this))
     } else {
       logger.warn(s"No storage driver defined for $storageType")
       None
@@ -83,7 +92,7 @@ case class StorageEntry(id: Option[Int], rootpath: Option[String], clientpath: O
       })
   }
 
-  def validatePathExists(filePath:String) = getStorageDriver.map(drv=>drv.pathExists(filePath:String)) match {
+  def validatePathExists(filePath:String, version:Int)(implicit mat:Materializer) = getStorageDriver.map(drv=>drv.pathExists(filePath, version)) match {
     case None=>Left(s"No storage driver exists for storage $id ($rootpath)!")
     case Some(result)=>Right(result)
   }
@@ -96,6 +105,7 @@ class StorageEntryRow(tag:Tag) extends Table[StorageEntry](tag, "StorageEntry") 
   )
 
   def id = column[Int]("id",O.PrimaryKey, O.AutoInc)
+  def nickname = column[Option[String]]("s_nickname")
   def rootpath = column[Option[String]]("s_root_path")
   def clientpath = column[Option[String]]("s_client_path")
   def storageType = column[String]("s_storage_type")
@@ -103,9 +113,11 @@ class StorageEntryRow(tag:Tag) extends Table[StorageEntry](tag, "StorageEntry") 
   def password = column[Option[String]]("s_password")
   def host = column[Option[String]]("s_host")
   def port = column[Option[Int]]("i_port")
+  def device = column[Option[String]]("s_device")
   def status = column[Option[StorageStatus.Value]]("e_status")
+  def supportsVersions = column[Boolean]("b_versions")
 
-  def * = (id.?,rootpath,clientpath,storageType,user,password,host,port,status) <> (StorageEntry.tupled, StorageEntry.unapply)
+  def * = (id.?,nickname, rootpath,clientpath,storageType,user,password,host,port,device, supportsVersions, status) <> (StorageEntry.tupled, StorageEntry.unapply)
 }
 
 
