@@ -2,12 +2,11 @@ package services
 
 import javax.inject.{Inject, Singleton}
 import akka.actor.{Actor, ActorRef, ActorSystem}
+import akka.stream.Materializer
 import models.{StorageEntry, StorageEntryHelper, StorageStatus}
 import play.api.{Configuration, Logger}
 import play.api.db.slick.DatabaseConfigProvider
 import slick.jdbc.PostgresProfile
-
-import scala.concurrent.duration._
 import scala.util.{Failure, Success}
 import scala.concurrent.ExecutionContext.Implicits.global
 
@@ -20,7 +19,7 @@ object StorageScanner {
 }
 
 @Singleton
-class StorageScanner @Inject() (dbConfigProvider:DatabaseConfigProvider, config:Configuration, actorSystem:ActorSystem) extends Actor {
+class StorageScanner @Inject() (dbConfigProvider:DatabaseConfigProvider, config:Configuration, actorSystem:ActorSystem)(implicit mat:Materializer) extends Actor {
   private val logger = Logger(getClass)
   import StorageScanner._
 
@@ -46,7 +45,7 @@ class StorageScanner @Inject() (dbConfigProvider:DatabaseConfigProvider, config:
         case Some(rootpath)=>
           entry.getStorageDriver match {
             case Some (storageDriver) =>
-              if (storageDriver.pathExists(rootpath)){
+              if (storageDriver.pathExists(rootpath, 0)){
                 logger.debug(s"Storage ${entry.storageType} ${entry.rootpath} is online")
                 if(entry.status.getOrElse(StorageStatus.UNKNOWN)!=StorageStatus.ONLINE) entry.copy(status=Some(StorageStatus.ONLINE)).save
               } else {
