@@ -17,7 +17,8 @@ class StorageMultistep extends CommonMultistepRoot {
             selectedType: 0,
             rootpath: "",
             clientpath: null,
-            loginDetails: {}
+            loginDetails: {},
+            enableVersions: false
         }
     }
 
@@ -38,9 +39,12 @@ class StorageMultistep extends CommonMultistepRoot {
                 loginDetails: {
                     hostname: response.data.result.host,
                     port: response.data.result.port ? parseInt(response.data.result.port) : null,
+                    device: response.data.result.device,
                     username: response.data.result.user,
                     password: response.data.result.password
-                }
+                },
+                enableVersions: response.data.result.supportsVersions,
+                nickname: response.data.result.nickname
             })
         }).catch(error=>{
             console.error(error);
@@ -67,7 +71,10 @@ class StorageMultistep extends CommonMultistepRoot {
                 component: <StorageTypeComponent strgTypes={this.state.strgTypes}
                                                  selectedType={this.state.selectedType}
                                                  currentStorage={this.props.currentEntry}
-                                                 valueWasSet={(type)=>this.setState({selectedType: type})}/>
+                                                 versionsAllowed={this.state.enableVersions}
+                                                 valueWasSet={(type)=>this.setState({selectedType: type})}
+                                                 versionsAllowedChanged={newValue=>this.setState({enableVersions: newValue})}
+                />
             },
             {
                 name: 'Login details',
@@ -95,6 +102,9 @@ class StorageMultistep extends CommonMultistepRoot {
                                                        rootpath={this.state.rootpath}
                                                        clientpath={this.state.clientpath}
                                                        currentEntry={this.state.currentEntry}
+                                                       enableVersions={this.state.enableVersions}
+                                                       nickname={this.state.nickname}
+                                                       nicknameChanged={evt=>this.setState({nickname: evt.target.value})}
                 />
             }
         ];
