@@ -12,6 +12,7 @@ class StorageLoginComponent extends React.Component {
         this.setState({
             hostname: this.props.loginDetails.hostname,
             port: this.props.loginDetails.port,
+            device: this.props.loginDetails.device,
             username: this.props.loginDetails.username,
             password: this.props.loginDetails.password
         })
@@ -23,6 +24,7 @@ class StorageLoginComponent extends React.Component {
         this.state = {
             hostname: "",
             port: "",
+            device: "",
             username: "",
             password: ""
         }
@@ -50,23 +52,27 @@ class StorageLoginComponent extends React.Component {
                         is not applicable, leave it blank
                     </p>
                     <form>
-                        <table>
+                        <table className="full-width">
                             <tbody>
                             <tr>
-                                <td>Storage host (or bucket name)</td>
-                                <td><input id="hostname_input" value={this.state.hostname} onChange={(event)=>this.setState({hostname: event.target.value})}/></td>
+                                <td className="narrow">Storage host (or bucket name)</td>
+                                <td><input className="full-width" id="hostname_input" value={this.state.hostname} onChange={(event)=>this.setState({hostname: event.target.value})}/></td>
                             </tr>
                             <tr>
-                                <td>Storage port (if applicable)</td>
-                                <td><input id="port_input" value={this.state.port} onChange={(event)=>this.setState({port: event.target.value})}/></td>
+                                <td className="narrow">Storage port (if applicable)</td>
+                                <td><input className="full-width" id="port_input" value={this.state.port} onChange={(event)=>this.setState({port: event.target.value})}/></td>
                             </tr>
                             <tr>
-                                <td>User name</td>
-                                <td><input id="username_input" value={this.state.username} onChange={(event)=>this.setState({username: event.target.value})}/></td>
+                                <td className="narrow">Device (if applicable)<p className="explanation">For ObjectMatrix storage, specify the "cluster ID" and "vault ID" separated by a comma</p></td>
+                                <td><input className="full-width" id="device_input" value={this.state.device} onChange={event=>this.setState({device: event.target.value})}/></td>
                             </tr>
                             <tr>
-                                <td>Password</td>
-                                <td><input id="password_input" type="password" value={this.state.password} onChange={(event)=>this.setState({password: event.target.value})}/></td>
+                                <td className="narrow">User name</td>
+                                <td><input className="full-width" id="username_input" value={this.state.username} onChange={(event)=>this.setState({username: event.target.value})}/></td>
+                            </tr>
+                            <tr>
+                                <td className="narrow">Password</td>
+                                <td><input className="full-width" id="password_input" type="password" value={this.state.password} onChange={(event)=>this.setState({password: event.target.value})}/></td>
                             </tr>
                             </tbody>
                         </table>
