@@ -40,7 +40,7 @@ class UpdateAdobeUuidSpec extends Specification with BuildMyApp {
       )), 10 seconds)
 
       val pe = futureResults.head.asInstanceOf[Try[ProjectEntry]].get
-      val pt = futureResults(1).asInstanceOf[Try[ProjectType]].get
+      val pt = futureResults(1).asInstanceOf[ProjectType]
 
       val result = Await.result(s.postrun("/tmp/test_update_uuid.prproj",pe,pt,dataCache,None,None),10 seconds)
       result must beSuccessfulTry
